Enable to keep the result of a DBus call packed
Sometimes, it might be better to keep the result of a DBus call
packed into a variant, because it can be a complicated data structure
and it is easier to modify it and send back to DBus.

Call the DBus method with an argument unpack_result=False to get
the packed result of the call.

Example:

  # Get an unpacked result.
  result = proxy.Method(arg1, args2)

  # Get a packed result.
  result = proxy.Method(arg1, args2, unpack_result=False)